testing/web-platform/tests/atomics/dedicated-worker-agent/wait/nan-timeout.html

<!doctype html>
<meta charset=utf-8>
<title></title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/resources/workerhelper.js></script>
<script>

// Copyright (C) 2015 Mozilla Corporation.  All rights reserved.
// This code is governed by the BSD license found in the LICENSE file.

/*---
description: >
  Test that Atomics.wait does not time out with a NaN timeout
---*/

window.addEventListener("load", function() {
  async_test(function(t) {
    var worker_code = `
      onmessage = function(e) {
        let ia = e.data[0];
        let ret = Atomics.wait(ia, 0, 0, NaN);
        postMessage([ret]);
      }
    `;
    var onmessage = this.step_func(function(e) {
      let val = e.data[0];
      assert_equals(val, "ok");
      t.done();
    });
    var ia = new Int32Array(new SharedArrayBuffer(Int32Array.BYTES_PER_ELEMENT));
    var worker = WorkerHelper.createFromString(worker_code, onmessage);
    worker.postMessage([ia]);
    window.setTimeout(() => Atomics.wake(ia, 0), 500);
  }, "atomics-dedicated-worker-wait-nan-timeout");
});

</script>
